Added multiple turrets

+++ Assets/Scripts/Enemy/EnemyAsset.cs
@@ -5,7 +5,7 @@ namespace Enemy
     [CreateAssetMenu(menuName = "Assets/Enemy Asset", fileName = "Enemy Asset")]
     public class EnemyAsset : ScriptableObject
     {
-        public int startHealth;
+        public float startHealth;
         public bool isFlyingEnemy;
 
         public EnemyView viewPrefab;
